final refactoring of terminal and gameplay
What was done. Was problem - in terminal characters was higher twice
than width. And before refactoring, gameplay make double change if
motion was horizontal. And therefore in gameplay was logic of display.
But in architect view this is not correct

And with this refactoring gameplay doesn't know about terminal moments at all